Change encoding tests to be unit tests
This allows us to revert the exposing of
rusk_proto. It also makes more sense
to have these tests as unit tests, since
we are not combining any parts of rusk
for these tests.
